requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.5
django-appconf==1.0.4
django-bootstrap-form==3.4
django-environ==0.4.5
django-extensions==2.2.9
django-filter==2.2.0
django-imagekit==4.0.2
djangorestframework==3.11.0
gunicorn==20.0.4
pilkit==2.0
Pillow==7.1.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
